Refactored Java event list fragment and info page
